ros_ws/src/see_camera_processing/src/ssds/requirements.txt
opencv-python
matplotlib
tensorboardX
torchvision
pyyaml
cython
pycocotools
